General code cleanup
Many changes:
- Replaces single quotes with double quotes for docstrings
- Using now new-style Python classes instead of old style
- Moved cache wrappers to their own module
- Sorted imports
- Removal of some extranous things (like parenthesis on a if)
