+++ playcert/lib/artist.py
@@ -1,21 +1,23 @@
-from pyechonest import config, artist as echonest_artist, song as echonest_song
 import sys
-import os
 import logging
 import urllib
-import requests
-import dill
 import collections
-from functools import wraps
+
+from pyechonest import config, artist as echonest_artist, song as echonest_song
+import os
+import requests
+
+from playcert.lib.cache_utils import cache_songs
 
 
 config.ECHO_NEST_API_KEY = os.environ['ECHONEST_KEY']
+
 log = logging.getLogger(__name__)
 
 Song = collections.namedtuple('Song', ['name', 'spotify_id'])
 
 
-class Artist:
+class Artist(object):
 
     def __repr__(self):
         return 'Artist(%s, %s)' % (self.name.encode('utf-8'), self.songs)
@@ -32,19 +34,18 @@ class Artist:
 
     @staticmethod
     def create_artist_from_text(text, venue, redis=None):
-        '''
+        """
         Static method that aims in extracting artist name from a text
         and creating an instance of this class based on it
-        '''
+        """
 
         # remove venue name from the title of the event
         # it will improve the finding of the artist name from the event title
         if venue:
-            text = text.replace(venue, "")
+            text = text.replace(venue, '')
 
         try:
-            event_artist = echonest_artist.extract(
-                text=text, results=1)
+            event_artist = echonest_artist.extract(text=text, results=1)
         except:
             log.error('could not find artist %s', sys.exc_info()[0])
             return
@@ -60,43 +61,11 @@ class Artist:
 
         return
 
-    def cache_songs(f):
-        '''
-        Caches the songs on redis
-        '''
-        @wraps(f)
-        def decorated_function(*args):
-            self = args[0]
-
-            if self.redis:
-                # try to get songs from cache and set them
-                songs = self.redis.hget('artist.songs', self.name)
-                songs = dill.loads(songs) if songs else ''
-
-                log.debug('Trying to obtain %s songs from cache', self.name)
-
-                if songs:
-                    self.songs = songs
-                    log.debug('%s songs obtained from cache', self.name)
-                    return
-
-            # find the songs
-            f(*args)
-
-            # save them on cache
-            if self.songs and self.redis:
-                # set songs on cache
-                self.redis.hset(
-                    'artist.songs', self.name, dill.dumps(self.songs))
-
-                log.debug('%s songs setted in cache', self.name)
-        return decorated_function
-
     @cache_songs
     def find_songs(self):
-        '''
-        finds songs for this artist
-        '''
+        """
+        Finds songs for this artist
+        """
 
         # first try in thisdayinmusic.net api
         try:
@@ -124,7 +93,6 @@ class Artist:
         if 'data' in artist_info:
             self.songs = [Song(song['name'], song['spotifyId'])
                           for song in artist_info['data']['tracks']['data']]
-
             return
 
         # couldn't find artist in thisdayinmusic, trying echonest
@@ -147,6 +115,4 @@ class Artist:
                 self.songs = tracks
 
         except:
-            log.error(
-                'could not find songs in echonest %s',
-                sys.exc_info()[0])
+            log.error('could not find songs in echonest %s', sys.exc_info()[0])
