+++ playcert/lib/event.py
@@ -1,12 +1,13 @@
-import artist
-import dill
 import logging
-from functools import wraps
+
+from playcert.lib.artist import Artist
+from playcert.lib.cache_utils import cache_artist
+
 
 log = logging.getLogger(__name__)
 
 
-class Event:
+class Event(object):
 
     def __init__(self, title, when, venue, artist_name=None, redis=None):
         self.title = title
@@ -19,45 +20,16 @@ class Event:
         if artist_name:
             log.debug(
                 'No need to find artist name, already have it from eventful %s', artist_name)
-            self.artist = artist.Artist(artist_name, self.redis)
+            self.artist = Artist(artist_name, self.redis)
         else:
             # finds artist from event title
             self.find_artist()
-
-    def cache_artist(f):
-        '''
-        Caches the artist on redis
-        '''
-        @wraps(f)
-        def decorated_function(*args):
-            self = args[0]
-
-            if self.redis:
-                artist_cache = self.redis.hget('text.artist', self.title)
-                artist_cache = dill.loads(artist_cache) if artist_cache else ''
-                if artist_cache:
-                    self.artist = artist_cache
-                    log.debug(
-                        'Obtained artist for %s from cache', self.title)
-                    return
-
-            f(*args)
-
-            if self.artist and self.redis:
-                # set artist on cache
-                self.redis.hset(
-                    'text.artist', self.title, dill.dumps(self.artist))
-                log.debug(
-                    'Setting artist (%s) for %s on cache', self.artist.name, self.title
-                )
-
-        return decorated_function
 
     @cache_artist
     def find_artist(self):
 
         # finds and creates artist from event title
-        self.artist = artist.Artist.create_artist_from_text(
+        self.artist = Artist.create_artist_from_text(
             self.title, self.venue, self.redis)
 
     def __repr__(self):
